requirements.txt
altair @ file:///tmp/build/80754af9/altair_1599835197802/work
anyio==3.6.2
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644553347904/work
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///croot/attrs_1668696182826/work
Babel==2.11.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blinker==1.4
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
branca==0.6.0
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==3.0.1
click @ file:///tmp/build/80754af9/click_1646056706450/work
comm @ file:///croot/comm_1671231121260/work
commonmark @ file:///Users/ktietz/demo/mc3/conda-bld/commonmark_1630649545323/work
cryptography @ file:///croot/cryptography_1673298753778/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.6.6
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908313000/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.16.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
folium==0.14.0
fonttools==4.25.0
fqdn==1.5.1
future @ file:///home/builder/ci_310/future_1640790123501/work
geographiclib==1.52
geopy==2.2.0
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///croot/gitpython_1674662789231/work
haversine==2.7.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562408262/work
ipykernel @ file:///croot/ipykernel_1671488378391/work
ipython @ file:///croot/ipython_1674681422581/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
jupyterlab_server==2.19.0
jupyterthemes==0.20.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lesscpy==0.15.1
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///opt/conda/conda-bld/matplotlib-suite_1660167928326/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.5.1
nbclient @ file:///tmp/build/80754af9/nbclient_1650290238894/work
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
notebook @ file:///croot/notebook_1668179881751/work
notebook_shim==0.2.2
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.5.2
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
platformdirs @ file:///opt/conda/conda-bld/platformdirs_1662711380096/work
plotly @ file:///tmp/abs_7afcdfad-dbbb-49d2-adea-186abf525c45jbnd8p95/croots/recipe/plotly_1658160053621/work
ply==3.11
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==8.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck @ file:///tmp/abs_c06301fd-7800-4748-8e9e-0f9ab34bc0cess8ywq6h/croots/recipe/pydeck_1657894564709/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
Pympler @ file:///tmp/build/80754af9/pympler_1602785470644/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
pyrsistent @ file:///home/builder/ci_310/pyrsistent_1640807196327/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz @ file:///croot/pytz_1671697431263/work
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich @ file:///tmp/abs_f016wesp06/croots/recipe/rich_1659708287203/work
semver @ file:///tmp/build/80754af9/semver_1603822362442/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
sniffio==1.3.0
soupsieve @ file:///croot/soupsieve_1666296392845/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.17.0
streamlit-folium==0.11.0
tenacity @ file:///tmp/abs_9ca5cd06-f53b-4ea6-8eea-954e11157fddo5mhurpp/croots/recipe/tenacity_1657899103026/work
terminado @ file:///croot/terminado_1671751832461/work
themes==0.0.4
tinycss2==1.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
tornado==6.2
traitlets==5.8.1
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzlocal @ file:///croot/tzlocal_1668627599277/work
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
validators @ file:///tmp/build/80754af9/validators_1612286467315/work
watchdog @ file:///home/builder/ci_310/watchdog_1640811339391/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354667/work
zipp @ file:///croot/zipp_1672387121353/work
